LayoutTests/http/tests/site-isolation/remote-frame-child-by-name.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script>
    if (window.testRunner) { testRunner.dumpAsText() }
</script>
<body>
<iframe src="http://localhost:8000/site-isolation/resources/remote-frame-child-by-name.html" name=childframename></iframe>
</body>
